make it possible to run pylint and fix linter warnings
also use lighterweight subprocess.check_output instead of the
real heavyweight execute_command_with_output for the multitude
of s3api get_object range requests
